+++ src/DaysView.js
@@ -16,9 +16,9 @@ var DateTimePickerDays = createClass({
 		tableChildren = [
 			React.createElement('thead', { key: 'th' }, [
 				React.createElement('tr', { key: 'h' }, [
-					React.createElement('th', { key: 'p', className: 'rdtPrev', onClick: this.props.subtractTime( 1, 'months' )}, React.createElement('span', {}, '‹' )),
+					React.createElement('th', { key: 'p', className: 'rdtPrev', onClick: this.props.navigate( -1, 'months' )}, React.createElement('span', {}, '‹' )),
 					React.createElement('th', { key: 's', className: 'rdtSwitch', onClick: this.props.showView( 'months' ), colSpan: 5, 'data-value': this.props.viewDate.month() }, locale.months( date ) + ' ' + date.year() ),
-					React.createElement('th', { key: 'n', className: 'rdtNext', onClick: this.props.addTime( 1, 'months' )}, React.createElement('span', {}, '›' ))
+					React.createElement('th', { key: 'n', className: 'rdtNext', onClick: this.props.navigate( 1, 'months' )}, React.createElement('span', {}, '›' ))
 				]),
 				React.createElement('tr', { key: 'd'}, this.getDaysOfWeek( locale ).map( function( day, index ) { return React.createElement('th', { key: day + index, className: 'dow'}, day ); }) )
 			]),
@@ -111,7 +111,7 @@ var DateTimePickerDays = createClass({
 	},
 
 	updateSelectedDate: function( event ) {
-		this.props.updateSelectedDate( event, true );
+		this.props.updateDate( event );
 	},
 
 	renderDay: function( props, currentDate ) {